rename app.css
Rough idea: can we isolate each UI thing to its own CSS/JS file?

This is handy for development, and might make it possible for a user
to selectively enable/disable treatments.

We'd eventually want to concatenate (and ideally minify) them all into
a single file to minimize impact of making many requests, but we
can ignore that until its in the 10-20 range.