change stats indexing
Progress on #68

Backfilling stats now happens in a worker thread while the system is
running. We make round-robin progress on each column in each table
in each DB.

We only index strings that seem likely to need autocomplete --
not numbers, not dates, not JSON values, not URLs.

JSON string values are still indexed, although now we also index
their row's pk.

We keep a max of 10 pks per value.

Still to come - update the autosuggest code to understand the new
schema.