scaffolding of cypress
progress on #84

I think my goal here is to write E2E smoke tests that
confirm you can do a series of actions, given a starting
point of a database in a specific, known-good configuration.

I'm checking in the dummy tests that Cypress creates for reference,
I'll delete them in a future commit.