view support for omnisearch
fix #92

This is kinda buggy - we're ultimately searching the underlying base
table, so we might suggest things that have 0 records in the view.

Worse, the view could be specific to the user's access, so we might
leak things in the base table that the user ought not have access to.

Let's see how this feels in practice, and maybe roll it back if it
seems confusing